cleaning stuff at the beginning of teamcity full script

+++ assembler/src/test/teamcity/debruijn_full.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 set -e
 pushd ../../../
-rm data/debruijn/ECOLI_IS220_QUAKE/K55/latest
-rm src/tools/quality/results/all.txt
+[ -e data/debruijn/ECOLI_IS220_QUAKE/K55/latest ] && rm data/debruijn/ECOLI_IS220_QUAKE/K55/latest
+[ -e src/tools/quality/results] && rm -rf src/tools/quality/results
 make clean
 ./cpcfg
 ./spades.py src/test/teamcity/spades_config.full.info
